move error marker out of virtual_branches mod
It's useful outside of it
