+++ crates/gitbutler-core/tests/suite/virtual_branches/mod.rs
@@ -1,7 +1,7 @@
 use std::path::PathBuf;
 use std::{fs, path, str::FromStr};
 
-use gitbutler_core::virtual_branches::errors::Marker;
+use gitbutler_core::error::Marker;
 use gitbutler_core::{
     git,
     projects::{self, Project, ProjectId},
